fix

See merge request crm/crm-api!178

+++ Framework/CRCIS.Web.INoor.CRM.Contract/Repositories/Cases/IImportCaseRepository.cs
@@ -13,7 +13,7 @@ namespace CRCIS.Web.INoor.CRM.Contract.Repositories.Cases
 {
     public interface IImportCaseRepository
     {
-        Task<DataResponse<int>> CreateAsync(ImportCaseCreateCommand command, List<int> caseSubjectIds);
+        Task<DataResponse<long>> CreateAsync(ImportCaseCreateCommand command, List<int> caseSubjectIds);
         Task<DataResponse<int>> CreateAndMoveToAdmin(ImportCaseCreateCommand command, int adminId, List<int> caseSubjectIds);
         Task<DataTableResponse<IEnumerable<ImportCaseGetFullDto>>> GetAsync(ImportCaseDataTableQuery query);
         Task<DataResponse<int>> AddCaseHistoryMoveCaseToCurrentAdminAsync(MoveCaseToCurrentAdminCardboardCommand command);
